+++ CHANGELOG.md
@@ -16,10 +16,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fix(k8sprocessor): store only necessary Pod data [#593][#593]
 
+### Changed
+
+- chore(deps): update OT core to v0.52.0 [#600]
+
 [Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.51.0-sumo-0...main
 [#590]: https://github.com/SumoLogic/sumologic-otel-collector/pull/590
 [#592]: https://github.com/SumoLogic/sumologic-otel-collector/pull/592
 [#593]: https://github.com/SumoLogic/sumologic-otel-collector/pull/593
+[#600]: https://github.com/SumoLogic/sumologic-otel-collector/pull/600
 
 ## [v0.51.0-sumo-0]